fix: use flexible deserializers for cat tool offset/limit parameters
Change offset and limit parameters to accept both integer and string
values (like "50" or 50), matching the pattern used by grep and ls tools.

Changes:
- Change offset type from i64 to isize to match ReadArgs pattern
- Add deserialize_flexible_isize_option for offset parameter
- Add deserialize_flexible_usize_option for limit parameter
- Update JSON schema to accept ["integer", "string", "null"] types

This fixes ToolCallError when AI agents pass string values instead
of integers for numeric parameters.

+++ backend/src/models/requests.rs
@@ -821,9 +821,9 @@ pub struct CatArgs {
     pub squeeze_blank: Option<bool>,
 
     // Line range filtering
-    #[serde(default)]
-    pub offset: Option<i64>,
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none", deserialize_with = "deserialize_flexible_isize_option")]
+    pub offset: Option<isize>,
+    #[serde(default, skip_serializing_if = "Option::is_none", deserialize_with = "deserialize_flexible_usize_option")]
     pub limit: Option<usize>,
 }
